+++ src/pocketmine/entity/Animal.php
@@ -24,7 +24,7 @@ declare(strict_types=1);
 namespace pocketmine\entity;
 
 
-use pocketmine\network\mcpe\protocol\types\EntityMetadataFlags;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataFlags;
 
 abstract class Animal extends Living implements Ageable{
