+++ src/pocketmine/entity/Entity.php
@@ -51,10 +51,10 @@ use pocketmine\network\mcpe\protocol\MoveActorAbsolutePacket;
 use pocketmine\network\mcpe\protocol\RemoveActorPacket;
 use pocketmine\network\mcpe\protocol\SetActorDataPacket;
 use pocketmine\network\mcpe\protocol\SetActorMotionPacket;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataCollection;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataFlags;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataProperties;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataTypes;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataCollection;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataFlags;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataProperties;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataTypes;
 use pocketmine\player\Player;
 use pocketmine\Server;
 use pocketmine\timings\Timings;
