+++ src/pocketmine/entity/Human.php
@@ -45,10 +45,10 @@ use pocketmine\network\mcpe\protocol\ActorEventPacket;
 use pocketmine\network\mcpe\protocol\AddPlayerPacket;
 use pocketmine\network\mcpe\protocol\PlayerListPacket;
 use pocketmine\network\mcpe\protocol\PlayerSkinPacket;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataProperties;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataTypes;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataProperties;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataTypes;
+use pocketmine\network\mcpe\protocol\types\entity\PlayerMetadataFlags;
 use pocketmine\network\mcpe\protocol\types\PlayerListEntry;
-use pocketmine\network\mcpe\protocol\types\PlayerMetadataFlags;
 use pocketmine\player\Player;
 use pocketmine\utils\UUID;
 use pocketmine\world\sound\TotemUseSound;
