+++ src/pocketmine/entity/Villager.php
@@ -24,9 +24,9 @@ declare(strict_types=1);
 namespace pocketmine\entity;
 
 use pocketmine\nbt\tag\CompoundTag;
-use pocketmine\network\mcpe\protocol\types\EntityLegacyIds;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataFlags;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataProperties;
+use pocketmine\network\mcpe\protocol\types\entity\EntityLegacyIds;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataFlags;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataProperties;
 
 class Villager extends Living implements Ageable{
 	public const PROFESSION_FARMER = 0;
