+++ src/pocketmine/entity/effect/EffectManager.php
@@ -26,7 +26,7 @@ namespace pocketmine\entity\effect;
 use pocketmine\entity\Living;
 use pocketmine\event\entity\EntityEffectAddEvent;
 use pocketmine\event\entity\EntityEffectRemoveEvent;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataProperties;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataProperties;
 use pocketmine\utils\Color;
 use function abs;
