+++ src/pocketmine/entity/projectile/Egg.php
@@ -25,7 +25,7 @@ namespace pocketmine\entity\projectile;
 
 use pocketmine\event\entity\ProjectileHitEvent;
 use pocketmine\item\VanillaItems;
-use pocketmine\network\mcpe\protocol\types\EntityLegacyIds;
+use pocketmine\network\mcpe\protocol\types\entity\EntityLegacyIds;
 use pocketmine\world\particle\ItemBreakParticle;
 
 class Egg extends Throwable{
