+++ src/pocketmine/network/mcpe/handler/InGamePacketHandler.php
@@ -75,11 +75,11 @@ use pocketmine\network\mcpe\protocol\ShowCreditsPacket;
 use pocketmine\network\mcpe\protocol\SpawnExperienceOrbPacket;
 use pocketmine\network\mcpe\protocol\SubClientLoginPacket;
 use pocketmine\network\mcpe\protocol\TextPacket;
-use pocketmine\network\mcpe\protocol\types\MismatchTransactionData;
-use pocketmine\network\mcpe\protocol\types\NormalTransactionData;
-use pocketmine\network\mcpe\protocol\types\ReleaseItemTransactionData;
-use pocketmine\network\mcpe\protocol\types\UseItemOnEntityTransactionData;
-use pocketmine\network\mcpe\protocol\types\UseItemTransactionData;
+use pocketmine\network\mcpe\protocol\types\inventory\MismatchTransactionData;
+use pocketmine\network\mcpe\protocol\types\inventory\NormalTransactionData;
+use pocketmine\network\mcpe\protocol\types\inventory\ReleaseItemTransactionData;
+use pocketmine\network\mcpe\protocol\types\inventory\UseItemOnEntityTransactionData;
+use pocketmine\network\mcpe\protocol\types\inventory\UseItemTransactionData;
 use pocketmine\network\mcpe\serializer\NetworkNbtSerializer;
 use pocketmine\player\Player;
 use function array_push;
