+++ src/pocketmine/network/mcpe/protocol/AddActorPacket.php
@@ -29,8 +29,8 @@ use pocketmine\entity\Attribute;
 use pocketmine\math\Vector3;
 use pocketmine\network\BadPacketException;
 use pocketmine\network\mcpe\handler\PacketHandler;
-use pocketmine\network\mcpe\protocol\types\EntityLegacyIds;
-use pocketmine\network\mcpe\protocol\types\EntityLink;
+use pocketmine\network\mcpe\protocol\types\entity\EntityLegacyIds;
+use pocketmine\network\mcpe\protocol\types\entity\EntityLink;
 use function array_search;
 use function count;
