--- src/pocketmine/network/mcpe/protocol/types/EntityMetadataTypes.php
@@ -1,41 +1,0 @@
-<?php
-
-/*
- *
- *  ____            _        _   __  __ _                  __  __ ____
- * |  _ \ ___   ___| | _____| |_|  \/  (_)_ __   ___      |  \/  |  _ \
- * | |_) / _ \ / __| |/ / _ \ __| |\/| | | '_ \ / _ \_____| |\/| | |_) |
- * |  __/ (_) | (__|   <  __/ |_| |  | | | | | |  __/_____| |  | |  __/
- * |_|   \___/ \___|_|\_\___|\__|_|  |_|_|_| |_|\___|     |_|  |_|_|
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * @author PocketMine Team
- * @link http://www.pocketmine.net/
- *
- *
-*/
-
-declare(strict_types=1);
-
-namespace pocketmine\network\mcpe\protocol\types;
-
-final class EntityMetadataTypes{
-
-	private function __construct(){
-		//NOOP
-	}
-
-	public const BYTE = 0;
-	public const SHORT = 1;
-	public const INT = 2;
-	public const FLOAT = 3;
-	public const STRING = 4;
-	public const SLOT = 5;
-	public const POS = 6;
-	public const LONG = 7;
-	public const VECTOR3F = 8;
-}
