+++ src/pocketmine/network/mcpe/protocol/types/command/CommandData.php
@@ -21,7 +21,10 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\network\mcpe\protocol\types;
+namespace pocketmine\network\mcpe\protocol\types\command;
+
+use pocketmine\network\mcpe\protocol\types\command\CommandEnum;
+use pocketmine\network\mcpe\protocol\types\command\CommandParameter;
 
 class CommandData{
 	/** @var string */
