+++ src/pocketmine/network/mcpe/protocol/types/entity/EntityMetadataProperties.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\network\mcpe\protocol\types;
+namespace pocketmine\network\mcpe\protocol\types\entity;
 
 final class EntityMetadataProperties{
