+++ src/pocketmine/network/mcpe/protocol/types/inventory/NetworkInventoryAction.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\network\mcpe\protocol\types;
+namespace pocketmine\network\mcpe\protocol\types\inventory;
 
 use pocketmine\inventory\AnvilInventory;
 use pocketmine\inventory\EnchantInventory;
