+++ src/pocketmine/network/mcpe/protocol/types/inventory/NormalTransactionData.php
@@ -21,30 +21,35 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\network\mcpe\protocol\types;
+namespace pocketmine\network\mcpe\protocol\types\inventory;
 
-use pocketmine\network\BadPacketException;
 use pocketmine\network\mcpe\protocol\InventoryTransactionPacket;
+use pocketmine\network\mcpe\protocol\types\inventory\TransactionData;
+use pocketmine\network\mcpe\protocol\types\inventory\NetworkInventoryAction;
 use pocketmine\network\mcpe\serializer\NetworkBinaryStream;
-use function count;
 
-class MismatchTransactionData extends TransactionData{
+class NormalTransactionData extends TransactionData{
 
 	public function getTypeId() : int{
-		return InventoryTransactionPacket::TYPE_MISMATCH;
+		return InventoryTransactionPacket::TYPE_NORMAL;
 	}
 
 	protected function decodeData(NetworkBinaryStream $stream) : void{
-		if(!empty($this->actions)){
-			throw new BadPacketException("Mismatch transaction type should not have any actions associated with it, but got " . count($this->actions));
-		}
+
 	}
 
 	protected function encodeData(NetworkBinaryStream $stream) : void{
 
 	}
 
-	public static function new() : self{
-		return new self; //no arguments
+	/**
+	 * @param NetworkInventoryAction[] $actions
+	 *
+	 * @return NormalTransactionData
+	 */
+	public static function new(array $actions) : self{
+		$result = new self();
+		$result->actions = $actions;
+		return $result;
 	}
 }
