+++ src/pocketmine/network/mcpe/protocol/types/inventory/UseItemOnEntityTransactionData.php
@@ -21,17 +21,20 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\network\mcpe\protocol\types;
+namespace pocketmine\network\mcpe\protocol\types\inventory;
 
 use pocketmine\item\Item;
 use pocketmine\math\Vector3;
 use pocketmine\network\mcpe\protocol\InventoryTransactionPacket;
+use pocketmine\network\mcpe\protocol\types\inventory\TransactionData;
 use pocketmine\network\mcpe\serializer\NetworkBinaryStream;
 
-class ReleaseItemTransactionData extends TransactionData{
-	public const ACTION_RELEASE = 0; //bow shoot
-	public const ACTION_CONSUME = 1; //eat food, drink potion
+class UseItemOnEntityTransactionData extends TransactionData{
+	public const ACTION_INTERACT = 0;
+	public const ACTION_ATTACK = 1;
 
+	/** @var int */
+	private $entityRuntimeId;
 	/** @var int */
 	private $actionType;
 	/** @var int */
@@ -39,7 +42,16 @@ class ReleaseItemTransactionData extends TransactionData{
 	/** @var Item */
 	private $itemInHand;
 	/** @var Vector3 */
-	private $headPos;
+	private $playerPos;
+	/** @var Vector3 */
+	private $clickPos;
+
+	/**
+	 * @return int
+	 */
+	public function getEntityRuntimeId() : int{
+		return $this->entityRuntimeId;
+	}
 
 	/**
 	 * @return int
@@ -65,35 +77,48 @@ class ReleaseItemTransactionData extends TransactionData{
 	/**
 	 * @return Vector3
 	 */
-	public function getHeadPos() : Vector3{
-		return $this->headPos;
+	public function getPlayerPos() : Vector3{
+		return $this->playerPos;
+	}
+
+	/**
+	 * @return Vector3
+	 */
+	public function getClickPos() : Vector3{
+		return $this->clickPos;
 	}
 
 	public function getTypeId() : int{
-		return InventoryTransactionPacket::TYPE_RELEASE_ITEM;
+		return InventoryTransactionPacket::TYPE_USE_ITEM_ON_ENTITY;
 	}
 
 	protected function decodeData(NetworkBinaryStream $stream) : void{
+		$this->entityRuntimeId = $stream->getEntityRuntimeId();
 		$this->actionType = $stream->getUnsignedVarInt();
 		$this->hotbarSlot = $stream->getVarInt();
 		$this->itemInHand = $stream->getSlot();
-		$this->headPos = $stream->getVector3();
+		$this->playerPos = $stream->getVector3();
+		$this->clickPos = $stream->getVector3();
 	}
 
 	protected function encodeData(NetworkBinaryStream $stream) : void{
+		$stream->putEntityRuntimeId($this->entityRuntimeId);
 		$stream->putUnsignedVarInt($this->actionType);
 		$stream->putVarInt($this->hotbarSlot);
 		$stream->putSlot($this->itemInHand);
-		$stream->putVector3($this->headPos);
+		$stream->putVector3($this->playerPos);
+		$stream->putVector3($this->clickPos);
 	}
 
-	public static function new(array $actions, int $actionType, int $hotbarSlot, Item $itemInHand, Vector3 $headPos) : self{
+	public static function new(array $actions, int $entityRuntimeId, int $actionType, int $hotbarSlot, Item $itemInHand, Vector3 $playerPos, Vector3 $clickPos) : self{
 		$result = new self;
 		$result->actions = $actions;
+		$result->entityRuntimeId = $entityRuntimeId;
 		$result->actionType = $actionType;
 		$result->hotbarSlot = $hotbarSlot;
 		$result->itemInHand = $itemInHand;
-		$result->headPos = $headPos;
+		$result->playerPos = $playerPos;
+		$result->clickPos = $clickPos;
 		return $result;
 	}
 }
