+++ src/pocketmine/network/mcpe/protocol/types/inventory/UseItemTransactionData.php
@@ -21,21 +21,25 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\network\mcpe\protocol\types;
+namespace pocketmine\network\mcpe\protocol\types\inventory;
 
 use pocketmine\item\Item;
 use pocketmine\math\Vector3;
 use pocketmine\network\mcpe\protocol\InventoryTransactionPacket;
+use pocketmine\network\mcpe\protocol\types\inventory\TransactionData;
 use pocketmine\network\mcpe\serializer\NetworkBinaryStream;
 
-class UseItemOnEntityTransactionData extends TransactionData{
-	public const ACTION_INTERACT = 0;
-	public const ACTION_ATTACK = 1;
+class UseItemTransactionData extends TransactionData{
+	public const ACTION_CLICK_BLOCK = 0;
+	public const ACTION_CLICK_AIR = 1;
+	public const ACTION_BREAK_BLOCK = 2;
 
 	/** @var int */
-	private $entityRuntimeId;
-	/** @var int */
 	private $actionType;
+	/** @var Vector3 */
+	private $blockPos;
+	/** @var int */
+	private $face;
 	/** @var int */
 	private $hotbarSlot;
 	/** @var Item */
@@ -44,19 +48,28 @@ class UseItemOnEntityTransactionData extends TransactionData{
 	private $playerPos;
 	/** @var Vector3 */
 	private $clickPos;
-
-	/**
-	 * @return int
-	 */
-	public function getEntityRuntimeId() : int{
-		return $this->entityRuntimeId;
-	}
+	/** @var int */
+	private $blockRuntimeId;
 
 	/**
 	 * @return int
 	 */
 	public function getActionType() : int{
 		return $this->actionType;
+	}
+
+	/**
+	 * @return Vector3
+	 */
+	public function getBlockPos() : Vector3{
+		return $this->blockPos;
+	}
+
+	/**
+	 * @return int
+	 */
+	public function getFace() : int{
+		return $this->face;
 	}
 
 	/**
@@ -87,37 +100,51 @@ class UseItemOnEntityTransactionData extends TransactionData{
 		return $this->clickPos;
 	}
 
+	/**
+	 * @return int
+	 */
+	public function getBlockRuntimeId() : int{
+		return $this->blockRuntimeId;
+	}
+
 	public function getTypeId() : int{
-		return InventoryTransactionPacket::TYPE_USE_ITEM_ON_ENTITY;
+		return InventoryTransactionPacket::TYPE_USE_ITEM;
 	}
 
 	protected function decodeData(NetworkBinaryStream $stream) : void{
-		$this->entityRuntimeId = $stream->getEntityRuntimeId();
 		$this->actionType = $stream->getUnsignedVarInt();
+		$this->blockPos = new Vector3();
+		$stream->getBlockPosition($this->blockPos->x, $this->blockPos->y, $this->blockPos->z);
+		$this->face = $stream->getVarInt();
 		$this->hotbarSlot = $stream->getVarInt();
 		$this->itemInHand = $stream->getSlot();
 		$this->playerPos = $stream->getVector3();
 		$this->clickPos = $stream->getVector3();
+		$this->blockRuntimeId = $stream->getUnsignedVarInt();
 	}
 
 	protected function encodeData(NetworkBinaryStream $stream) : void{
-		$stream->putEntityRuntimeId($this->entityRuntimeId);
 		$stream->putUnsignedVarInt($this->actionType);
+		$stream->putBlockPosition($this->blockPos->x, $this->blockPos->y, $this->blockPos->z);
+		$stream->putVarInt($this->face);
 		$stream->putVarInt($this->hotbarSlot);
 		$stream->putSlot($this->itemInHand);
 		$stream->putVector3($this->playerPos);
 		$stream->putVector3($this->clickPos);
+		$stream->putUnsignedVarInt($this->blockRuntimeId);
 	}
 
-	public static function new(array $actions, int $entityRuntimeId, int $actionType, int $hotbarSlot, Item $itemInHand, Vector3 $playerPos, Vector3 $clickPos) : self{
+	public static function new(array $actions, int $actionType, Vector3 $blockPos, int $face, int $hotbarSlot, Item $itemInHand, Vector3 $playerPos, Vector3 $clickPos, int $blockRuntimeId) : self{
 		$result = new self;
 		$result->actions = $actions;
-		$result->entityRuntimeId = $entityRuntimeId;
 		$result->actionType = $actionType;
+		$result->blockPos = $blockPos;
+		$result->face = $face;
 		$result->hotbarSlot = $hotbarSlot;
 		$result->itemInHand = $itemInHand;
 		$result->playerPos = $playerPos;
 		$result->clickPos = $clickPos;
+		$result->blockRuntimeId = $blockRuntimeId;
 		return $result;
 	}
 }
