+++ src/pocketmine/network/mcpe/protocol/types/inventory/WindowTypes.php
@@ -22,7 +22,7 @@
 declare(strict_types=1);
 
 
-namespace pocketmine\network\mcpe\protocol\types;
+namespace pocketmine\network\mcpe\protocol\types\inventory;
 
 
 interface WindowTypes{
