+++ src/pocketmine/world/particle/FloatingTextParticle.php
@@ -30,9 +30,9 @@ use pocketmine\math\Vector3;
 use pocketmine\network\mcpe\protocol\AddPlayerPacket;
 use pocketmine\network\mcpe\protocol\PlayerListPacket;
 use pocketmine\network\mcpe\protocol\RemoveActorPacket;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataFlags;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataProperties;
-use pocketmine\network\mcpe\protocol\types\EntityMetadataTypes;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataFlags;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataProperties;
+use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataTypes;
 use pocketmine\network\mcpe\protocol\types\PlayerListEntry;
 use pocketmine\utils\UUID;
 use function str_repeat;
